moved logger helpers to common_logger_helpers.hpp file and refactored all lambda simulator formatter as static methods of cadmium::logger::simulator_formator.
