+++ include/cadmium/engine/pdevs_simulator.hpp
@@ -44,6 +44,8 @@ namespace cadmium {
         template<template<typename T> class MODEL, typename TIME, typename LOGGER>
         //TODO: implement the debugging
         class simulator {
+            using formatter=typename cadmium::logger::simulator_formatter<MODEL, TIME>;
+
             using input_ports=typename MODEL<TIME>::input_ports;
             using input_bags=typename make_message_bags<input_ports>::type;
             using output_ports=typename MODEL<TIME>::output_ports;
@@ -74,33 +76,13 @@ namespace cadmium {
              * @param initial_time is the start time
              */
             void init(TIME initial_time) {
-                auto log_info_init = [](TIME t) -> std::string {
-                     std::ostringstream oss;
-                     oss << "Simulator for model ";
-                     oss << boost::typeindex::type_id<MODEL<TIME>>().pretty_name();
-                     oss << " initialized to time ";
-                     oss << t;
-                     return oss.str();
-                 };
-                LOGGER::template log<cadmium::logger::logger_info, decltype(log_info_init), TIME>(log_info_init, initial_time);
-
+                LOGGER::template log<cadmium::logger::logger_info, decltype(formatter::log_info_init), TIME>(formatter::log_info_init, initial_time);
 
                 _last=initial_time;
                 cadmium::concept::atomic_model_assert<MODEL>();
                 _next = initial_time + _model.time_advance();
                 
-                
-                auto log_state = [](const typename model_type::state_type& s) -> std::string {
-                    std::ostringstream oss;
-                    oss << "State for model ";
-                    oss << boost::typeindex::type_id<model_type>().pretty_name();
-                    oss << " is ";
-                    oss << s;
-                    return oss.str();
-                };
-                
-                LOGGER::template log<cadmium::logger::logger_state, decltype(log_state), const typename model_type::state_type&>(log_state, _model.state);
-
+                LOGGER::template log<cadmium::logger::logger_state, decltype(formatter::log_state), const typename model_type::state_type&>(formatter::log_state, _model.state);
             }
 
 
@@ -109,15 +91,7 @@ namespace cadmium {
             }
 
             void collect_outputs(const TIME &t) {
-                auto log_info_collect = [](TIME t) -> std::string {
-                     std::ostringstream oss;
-                     oss << "Simulator for model ";
-                     oss << boost::typeindex::type_id<model_type>().pretty_name();
-                     oss << " collecting output at time ";
-                     oss << t;
-                     return oss.str();
-                };
-                LOGGER::template log<cadmium::logger::logger_info, decltype(log_info_collect), TIME>(log_info_collect, t);
+                LOGGER::template log<cadmium::logger::logger_info, decltype(formatter::log_info_collect), TIME>(formatter::log_info_collect, t);
 
                 //cleanning the inbox and producing outbox
                 _inbox = in_bags_type{};
@@ -131,15 +105,7 @@ namespace cadmium {
                     _outbox = out_bags_type();
                 }
 
-                auto log_messages_collect = [](const out_bags_type& ob) -> std::string {
-                     std::ostringstream oss;
-                     print_messages_by_port(oss, ob);
-                     oss << " generated by model ";
-                     oss << boost::typeindex::type_id<model_type>().pretty_name();
-                     return oss.str();
-                };
-                LOGGER::template log<cadmium::logger::logger_messages, decltype(log_messages_collect), out_bags_type>(log_messages_collect, _outbox);
-
+                LOGGER::template log<cadmium::logger::logger_messages, decltype(formatter::log_messages_collect), out_bags_type>(formatter::log_messages_collect, _outbox);
             }
 
             /**
@@ -165,28 +131,8 @@ namespace cadmium {
                 //clean outbox because messages are routed before calling this funtion at a higher level
                 _outbox = out_bags_type{};
 
-                auto log_info_advance = [](const TIME& from, const TIME& to) -> std::string {
-                     std::ostringstream oss;
-                     oss << "Simulator for model ";
-                     oss << boost::typeindex::type_id<model_type>().pretty_name();
-                     oss << " advancing simulation from time ";
-                     oss << from;
-                     oss << " to ";
-                     oss << to;
-                     return oss.str();
-                };
-                LOGGER::template log<cadmium::logger::logger_info, decltype(log_info_advance), TIME>(log_info_advance, _last, t);
-
-                auto log_local_time = [](const TIME& from, const TIME& to) -> std::string {
-                    std::ostringstream oss;
-                    oss << "Elapsed in model ";
-                    oss << boost::typeindex::type_id<model_type>().pretty_name();
-                    oss << " is ";
-                    oss << (to - from);
-                    oss << "s";
-                    return oss.str();
-                };
-                LOGGER::template log<cadmium::logger::logger_local_time, decltype(log_local_time), TIME>(log_local_time, _last, t);
+                LOGGER::template log<cadmium::logger::logger_info, decltype(formatter::log_info_advance), TIME>(formatter::log_info_advance, _last, t);
+                LOGGER::template log<cadmium::logger::logger_local_time, decltype(formatter::log_local_time), TIME>(formatter::log_local_time, _last, t);
 
                 if (t < _last) {
                     throw std::domain_error("Event received for executing in the past of current simulation time");
@@ -217,23 +163,16 @@ namespace cadmium {
                     }
                 }
 
-                auto log_state = [](const typename model_type::state_type& s) -> std::string {
-                    std::ostringstream oss;
-                    oss << "State for model ";
-                    oss << boost::typeindex::type_id<model_type>().pretty_name();
-                    oss << " is ";
-                    oss << s;
-                    return oss.str();
-                };
-
-                LOGGER::template log<cadmium::logger::logger_state, decltype(log_state), const typename model_type::state_type&>(log_state, _model.state);
+                LOGGER::template log<cadmium::logger::logger_state, decltype(formatter::log_state), const typename model_type::state_type&>(formatter::log_state, _model.state);
             }
     //TODO: use enable_if functions to give access to read state and messages in debug mode
         };
 
 
         template<template<typename T> class MODEL, typename TIME, typename LOGGER>
-        class simulator_dynamic {
+        class dynamic_simulator {
+            using formatter=typename cadmium::logger::simulator_formatter<MODEL, TIME>;
+
             using input_ports=typename modeling::dynamic_atomic<MODEL, TIME>::input_ports;
             using output_ports=typename modeling::dynamic_atomic<MODEL, TIME>::output_ports;
             using in_bags_type=typename make_message_bags<input_ports>::type;
@@ -249,18 +188,17 @@ namespace cadmium {
             using model_type=MODEL<TIME>;
 
             /**
-             * @brief simulator constructs by default
-             */
-//            simulator() {}
-
-            /**
              * @brief constructor is used as init function, sets the start time
              * @param initial_time is the start time
              */
             void init(TIME initial_time) {
+                LOGGER::template log<cadmium::logger::logger_info, decltype(formatter::log_info_init), TIME>(formatter::log_info_init, initial_time);
+
                 _last=initial_time;
                 cadmium::concept::atomic_model_assert<MODEL>();
                 _next = initial_time + _model.time_advance();
+
+                LOGGER::template log<cadmium::logger::logger_state, decltype(formatter::log_state), const typename model_type::state_type&>(formatter::log_state, _model.state);
             }
 
             TIME next() const noexcept{
@@ -268,6 +206,7 @@ namespace cadmium {
             }
 
             void collect_outputs(const TIME &t) {
+                LOGGER::template log<cadmium::logger::logger_info, decltype(formatter::log_info_collect), TIME>(formatter::log_info_collect, t);
 
                 //cleanning the inbox and producing outbox
                 _inbox = dynamic_message_bags();
@@ -279,6 +218,8 @@ namespace cadmium {
                 } else {
                     _outbox = dynamic_message_bags();
                 }
+
+                LOGGER::template log<cadmium::logger::logger_messages, decltype(formatter::log_messages_collect), out_bags_type>(formatter::log_messages_collect, _outbox);
             }
 
             /**
@@ -303,6 +244,9 @@ namespace cadmium {
             void advance_simulation(TIME t) {
                 //clean outbox because messages are routed before calling this funtion at a higher level
                 _outbox = dynamic_message_bags();
+
+                LOGGER::template log<cadmium::logger::logger_info, decltype(formatter::log_info_advance), TIME>(formatter::log_info_advance, _last, t);
+                LOGGER::template log<cadmium::logger::logger_local_time, decltype(formatter::log_local_time), TIME>(formatter::log_local_time, _last, t);
 
                 if (t < _last) {
                     throw std::domain_error("Event received for executing in the past of current simulation time");
@@ -332,6 +276,8 @@ namespace cadmium {
                         }
                     }
                 }
+
+                LOGGER::template log<cadmium::logger::logger_state, decltype(formatter::log_state), const typename model_type::state_type&>(formatter::log_state, _model.state);
             }
         };
