+++ include/cadmium/logger/common_loggers.hpp
@@ -28,6 +28,7 @@
 #define COMMON_LOGGERS_HPP
 
 #include <cadmium/logger/logger.hpp>
+#include <cadmium/logger/common_loggers_helpers.hpp>
 
 namespace cadmium {
     namespace logger {    
@@ -98,6 +99,73 @@ namespace cadmium {
         //a logger that should not match any source, mostly for testing and debug purposes
         struct not_matching_source :public logger_source{};
         using not_logger=logger<not_matching_source, verbatim_formatter, cout_sink_provider>;
+
+        template <template<typename T> class MODEL, typename TIME>
+        struct simulator_formatter {
+
+            using model_type=MODEL<TIME>;
+            using input_ports=typename model_type::input_ports;
+            using output_ports=typename model_type::output_ports;
+            using in_bags_type=typename make_message_bags<input_ports>::type;
+            using out_bags_type=typename make_message_bags<output_ports>::type;
+
+            static std::string log_info_init(TIME t) {
+                std::ostringstream oss;
+                oss << "Simulator for model ";
+                oss << boost::typeindex::type_id<model_type>().pretty_name();
+                oss << " initialized to time ";
+                oss << t;
+                return oss.str();
+            }
+
+            static std::string log_state(const typename model_type::state_type& s) {
+                std::ostringstream oss;
+                oss << "State for model ";
+                oss << boost::typeindex::type_id<model_type>().pretty_name();
+                oss << " is ";
+                oss << s;
+                return oss.str();
+            };
+
+            static std::string log_info_collect(TIME t) {
+                std::ostringstream oss;
+                oss << "Simulator for model ";
+                oss << boost::typeindex::type_id<model_type>().pretty_name();
+                oss << " collecting output at time ";
+                oss << t;
+                return oss.str();
+            };
+
+            static std::string log_messages_collect(const out_bags_type& ob) {
+                std::ostringstream oss;
+                print_messages_by_port(oss, ob);
+                oss << " generated by model ";
+                oss << boost::typeindex::type_id<model_type>().pretty_name();
+                return oss.str();
+            };
+
+            static std::string log_info_advance(const TIME& from, const TIME& to) {
+                std::ostringstream oss;
+                oss << "Simulator for model ";
+                oss << boost::typeindex::type_id<model_type>().pretty_name();
+                oss << " advancing simulation from time ";
+                oss << from;
+                oss << " to ";
+                oss << to;
+                return oss.str();
+            };
+
+            static std::string log_local_time(const TIME& from, const TIME& to) {
+                std::ostringstream oss;
+                oss << "Elapsed in model ";
+                oss << boost::typeindex::type_id<model_type>().pretty_name();
+                oss << " is ";
+                oss << (to - from);
+                oss << "s";
+                return oss.str();
+            };
+        };
+
     }
 }
